+++ tests/negated-content.test.js
@@ -1,0 +1,26 @@
+import * as path from 'path'
+import { run, css } from './util/run'
+
+it('should be possible to use negated content patterns', () => {
+  let config = {
+    content: [
+      path.resolve(__dirname, './negated-content-*.test.html'),
+      '!' + path.resolve(__dirname, './negated-content-ignore.test.html'),
+    ],
+    corePlugins: { preflight: false },
+  }
+
+  let input = css`
+    @tailwind base;
+    @tailwind components;
+    @tailwind utilities;
+  `
+
+  return run(input, config).then((result) => {
+    expect(result.css).toMatchFormattedCss(css`
+      .uppercase {
+        text-transform: uppercase;
+      }
+    `)
+  })
+})